# src/backend/doc/swagger.yaml
swagger: "2.0"
info:
  description: "TestBuDDy is a test management, requirement management, automation friendly tool."
  version: "1.0.0"
  title: "TestBuDDy"
  contact:
    email: "barbora.blozonova@gmail.com"
  license:
    name: "MIT licence"
host: "127.0.0.1:5000"
basePath: "/"
...
schemes:
- "http"
paths:
...
  /init-projects:
    post:
      tags:
      - "Demo Initialization"
      summary: "Initialize projects"
      description: "(table: project)"
      produces:
      - "text/html"
      responses:
        200:
          description: "Operation successful. Test projects initialized."
        500:
          description: "Operation failed."
...
  /projects/{proj_id}:
    get:
      tags:
      - "Project management"
      summary: "Get project (by set path ID)"
      description: ""
      produces:
      - "application/json"
      - "text/html"
      parameters:
      - name: "proj_id"
        in: "path"
        type: "integer"
        required: true
      responses:
        404:
          description: "Project not found."
        200:
          schema:
            type: "object"
            properties:
              message:
                type: "string"
              object:
                  $ref: "#/definitions/Project"
          description: "Operation successful. Returns project in JSON response."
    put:
      tags:
      - "Project management"
      summary: "Update project"
      description: ""
      produces:
      - "application/json"
      - "text/html"
      parameters:
      - name: "proj_id"
        in: "path"
        type: "integer"
        required: true
      - name: "project info"
        in: "body"
        description: "Project information"
        required: true
        schema:
          type: "object"
          properties:
            name:
              type: "string"
            language_processor:
              type: "string"
            ci_params:
              type: "object"
              description: "required for Gitlab CI integration"
              required:
              - key
              properties:
                key:
                  type: "string"
            issue_tracker:
              type: "string"
            issue_tracker_params:
              type: "object"
              description: "required for JIRA integration"
              required:
              - server
              - username
              - password
              - project_shortcut
              properties:
                server:
                  type: "string"
                username:
                  type: "string"
                password:
                  type: "string"
                project_shortcut:
                  type: "string"
      responses:
        200:
          description: "Operation successful. Returns updated project in JSON response."
          schema:
            type: "object"
            properties:
              message:
                type: "string"
              object:
                  $ref: "#/definitions/Project"
        400:
          description: "Invalid body."
        404:
          description: "Project not found."
        415:
          description: "Unsupported media type."
        500:
          description: "Operation failed."
    delete:
      tags:
      - "Project management"
      summary: "Clean project repository and delete project"
      description: ""
      produces:
      - "text/html"
      parameters:
      - name: "proj_id"
        in: "path"
        type: "integer"
        required: true
      - name: "delete_project_in_dbs"
        description: "Delete project in database."
        in: "query"
        type: "boolean"
        default: false
      responses:
        200:
          description: "Operation successful."
        404:
          description: "Project not found."
        500:
          description: "Operation failed."
...
definitions:
...
  Project:
    type: "object"
    required:
    - id
    - name
    - repo_url
    - language_processor
    - ci_communicator
    - ci_params
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      repo_url:
        type: "string"
      language_processor:
        type: "string"
      ci_communicator:
        type: "string"
      ci_params:
        type: "object"
        description: "required for Gitlab CI integration"
        required:
        - key
        properties:
          key:
            type: "string"
          proj_path:
            type: "string"
          server_base:
            type: "string"
          server_url:
            type: "string"
      issue_tracker:
        type: "string"
      issue_tracker_params:
        type: "object"
        description: "required for JIRA integration"
        required:
        - server
        - username
        - password
        - project_shortcut
        properties:
          server:
            type: "string"
          username:
            type: "string"
          password:
            type: "string"
          project_shortcut:
            type: "string"
...